+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [0.21.0](https://kaos.sh/pachca/0.21.0)
 
 - Improved method `UploadFile` to pipe data instead of buffering
+- Added helper `GetLatestMessages`
 
 ### [0.20.2](https://kaos.sh/pachca/0.20.2)